Move checkpoint save handlers to separate module
