+++ frogbox/pipelines/pipeline.py
@@ -1,20 +1,16 @@
 from typing import (
-    Callable,
     Union,
     Optional,
     Any,
-    Dict,
+    Callable,
     Sequence,
-    Mapping,
     Tuple,
+    Dict,
 )
 from abc import ABC, abstractmethod
-import os
 from os import PathLike
 from pathlib import Path
 import datetime
-import tempfile
-import stat
 import torch
 from torch.utils.data import Dataset, DataLoader
 from ignite.engine import Engine, Events, CallableEventWithFilter
@@ -22,6 +18,7 @@ from ignite.handlers import global_step_from_engine, Checkpoint
 from ignite.handlers.checkpoint import BaseSaveHandler
 from ignite.handlers.base_logger import BaseLogger
 from accelerate import Accelerator
+from .save_handler import NoneSaveHandler, AccelerateDiskSaver
 from ..config import (
     Config,
     CheckpointMode,
@@ -29,87 +26,6 @@ from ..config import (
     create_object_from_config,
     parse_log_interval,
 )
-
-
-class NoneSaveHandler(BaseSaveHandler):
-    """@private"""
-
-    def __call__(
-        self,
-        checkpoint: Mapping,
-        filename: str,
-        metadata: Optional[Mapping] = None,
-    ) -> None:
-        pass
-
-    def remove(self, filename: str) -> None:
-        pass
-
-
-class AccelerateDiskSaver(BaseSaveHandler):
-    """@private"""
-
-    def __init__(
-        self,
-        dirname: Union[str, PathLike],
-        accelerator: Accelerator,
-        to_unwrap: Optional[Sequence[str]] = None,
-        atomic: bool = True,
-        **kwargs,
-    ):
-        self.dirname = Path(dirname).expanduser()
-        self.accelerator = accelerator
-        self.to_unwrap = to_unwrap
-        self.atomic = atomic
-        self.kwargs = kwargs
-
-        if not self.dirname.exists():
-            self.dirname.mkdir(parents=True)
-
-    def __call__(
-        self,
-        checkpoint: Mapping,
-        filename: str,
-        metadata: Optional[Mapping] = None,
-    ) -> None:
-        to_unwrap = self.to_unwrap if self.to_unwrap else []
-        unwrapped_checkpoint = {}
-        for key in checkpoint:
-            unwrapped_checkpoint[key] = (
-                self.accelerator.unwrap_model(checkpoint[key])
-                if key in to_unwrap
-                else checkpoint[key]
-            )
-
-        path = self.dirname / filename
-        self._save_func(unwrapped_checkpoint, path, self.accelerator.save)
-
-    def _save_func(
-        self, checkpoint: Mapping, path: Path, func: Callable
-    ) -> None:
-        if not self.atomic:
-            func(checkpoint, path, **self.kwargs)
-        else:
-            tmp = tempfile.NamedTemporaryFile(delete=False, dir=self.dirname)
-            tmp_file = tmp.file
-            tmp_name = tmp.name
-            try:
-                func(checkpoint, tmp_file, **self.kwargs)
-            except BaseException:
-                tmp.close()
-                os.remove(tmp_name)
-                raise
-            else:
-                tmp.close()
-                os.replace(tmp.name, path)
-                # append group/others read mode
-                os.chmod(
-                    path, os.stat(path).st_mode | stat.S_IRGRP | stat.S_IROTH
-                )
-
-    def remove(self, filename: str) -> None:
-        path = self.dirname / filename
-        path.unlink()
 
 
 class Pipeline(ABC):
